+++ tests/Logitar.EventSourcing.UnitTests/AggregateIdTests.cs
@@ -10,6 +10,11 @@ public class AggregateIdTests
   {
     AggregateId id = new();
     Assert.Equal(string.Empty.GetHashCode(), id.GetHashCode());
+
+    TestEvent e = new();
+    Assert.Equal(string.Empty, e.AggregateId.ToString());
+    Assert.Equal(id, e.AggregateId);
+    Assert.Equal(string.Empty.GetHashCode(), e.AggregateId.GetHashCode());
   }
 
   [Theory]