+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-Nothing yet.
+### Fixed
+
+- Fixed yet again default AggregateId bugs.
 
 ## [2.0.0] 2023-02-22